feat: UserService 추상화, Memory, Redis 분리
